Paddle/requirements.txt
numpy                              1.19.5
matplotlib                         3.3.4
opencv-python                      4.1.0.25
packaging                          16.8
paddlepaddle-gpu                   2.2.1
pandas                             1.1.5
Pillow                             8.4.0
scikit-image                       0.13.1
scikit-learn                       0.19.1
scipy                              1.5.4
sympy                              1.1.1
tensorboard                        1.14.0
tensorboardX                       1.6
torch                              1.10.0
torchvision                        0.11.1
tqdm                               4.32.1
x2paddle                           1.3.6
